decapitalising source file

+++ test/world.jl
@@ -6,7 +6,7 @@ myspace = (GraphSpace(SimpleGraph(10,10)),RealSpace{1,Float64}())
 myagents = [Agent(myspace,ancestors=true,rates=true) for i in 1:10]
 d(X,Y,t) = gaussian(X[1],Y[1],1)
 b(X,Y,t) = gaussian(X[1],0,1)
-D = [Int16(1),Float64(1.)]
+D = [Float32(1),Float64(1.)]
 mu = [1.,1.]
 NMax = 100
 p = Dict{String,Any}();@pack! p = d,b,D,mu,NMax
@@ -17,7 +17,7 @@ p = Dict{String,Any}();@pack! p = d,b,D,mu,NMax
 @test typeof(myagents) <: Vector{A} where {A<:AbstractAgentM}
 
 w = EvoId.World(myagents,myspace,p)
-@test typeof(w.p["D"][1]) == Int64
+@test typeof(w.p["D"][1]) == Float32
 @test size(w) ≈ 10
 newa = give_birth(1,w)
 addAgent!(w,newa)
